requirements.txt
streamlit==1.21.0
joblib==1.2.0
re
pdfminer.six==20201018
pytesseract==0.3.10
pdf2image==1.16.3
fpdf==1.7.2
scikit-learn==1.2.2



